Move numba_.py to core/util
